pyproject.toml
[project]
name = "the-clinic"
version = "0.1.0"
description = "University of Chicago Data Science Clinic website"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "graphviz>=0.20.1",
    "networkx>=3.0",
    "matplotlib>=3.5.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = [
  # by default, ruff enables all F and a subset of E rules
  # for a description of rulesets and examples, see:
  # https://docs.astral.sh/ruff/rules/
  "B",      # bugbear - identifies likely bugs
  "C4",     # flake8-comprehensions - identifies redundant calls in comprehensions
  "D",      # pydocstyle - requires docstrings on all public modules/functions/classes
  "I",      # isort
  "N801",   # pep8-naming - enforce ClassNamesStyle
  "N804",   # pep8-naming - first arg of classmethod should be cls
  "N805",   # pep8-naming - first arg of method should be self
  "PD",     # pandas-vet
  "PLR2004",# Pylint no magic values!
  "PTH",    # prefer pathlib over os.path as seems to be the consensus
  "Q",      # flake8-quotes - double quotes preferred (like black)
  "S",      # flake8-bandit - security checks
  "UP",     # pyupgrade - requires preferred syntax in newer python versions
  "YTT",    # flake8-2020 - corrects code that doesn't anticipate 2 digit python minor version  
]

ignore = [
  "D415",   # allow first line of doctstrings to end without punctuation
  "B905",    # won't work on python <3.10
  "S101"    # Permit assertions
]

[tool.ruff.lint.pydocstyle]
convention = "google"
